requirements.txt
flask
torch
torchvision
timm
segmentation-models-pytorch
albumentations
opencv-python-headless
pillow
matplotlib
